Extract wiki model name parser.
